Move player indicator to a separate class
